requirements.txt
Django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
djangorestframework
django-cors-headers